build script robustness

+++ scripts/pkgBundleTypes.cjs
@@ -9,29 +9,34 @@ bundlePkgTypes(process.cwd())
 
 async function bundlePkgTypes(dir) {
   const pkgConfig = getPkgConfig(dir)
-  const { exportPaths } = analyzePkgConfig(pkgConfig)
+  const { exportSubnames, exportPaths } = analyzePkgConfig(pkgConfig)
   await live(['rollup', '--config', rollupConfigPath])
 
   shell.rm('tsconfig.tsbuildinfo') // tsbuild cache is invalid now
   shell.cd('dist')
-  shell.rm('impl.d.ts.map') // rollup can't product a map, old map is useless
+
+  // rollup can't product a map, old map is useless
+  if (exportSubnames.includes('impl')) {
+    shell.rm('impl.d.ts.map')
+  }
 
   // generate filenames without directory/extension
-  const simpleExportPaths = exportPaths.reduce((accum, exportPath) => {
-    accum[path.basename(exportPath).replace(/\..*/, '')] = true
+  const exportDistNames = exportPaths.reduce((accum, exportPath) => {
+    accum[removeExt(path.basename(exportPath))] = true
     return accum
   }, {})
 
   // iterate all top-level files in the 'dist' directory
   // remove anything that's not related to a package export
   for (const filename of shell.ls()) {
-    const match = filename.match(/^([^.]*)\./) // everything before first dot
-    if (match) {
-      if (!simpleExportPaths[match[1]]) { // not listed in package exports
-        shell.rm(filename)
-      }
-    } else { // a directory
+    if (filename.indexOf('.') === -1) { // directory
       shell.rm('-rf', filename)
+    } else if (!exportDistNames[removeExt(filename)]) { // not listed in package exports
+      shell.rm(filename)
     }
   }
+}
+
+function removeExt(filename) {
+  return filename.replace(/\.(js|cjs|ts|d\.ts)(\.map)?$/, '')
 }
